Refactor file structure to enable separation of listener from testable components
