+++ package.json
@@ -5,7 +5,8 @@
   "main": "server.js",
   "scripts": {
     "start": "node server.js",
-    "test": "jest"
+    "test": "jest",
+    "check-handles": "jest --detectOpenHandles"
   },
   "author": "Nat Poulson",
   "license": "ISC",
