+++ server.js
@@ -1,12 +1,6 @@
 // TODO: Implement express.js and connect middleware
-const express = require('express');
-
-const app = express();
+const app = require('./app');
 
 const PORT = process.env.PORT || 3001;
 
-app.use(express.static('public'));
-
-app.listen(PORT, () => {
-    console.log(`Now listening on http://localhost:${PORT}`);
-});
+app.listen(PORT);
